259 - Update View Items page
- Update view selected item dialog page to support
max width of sm.
- Update ImagePicker component to support uploading
image from the EditInventory page.
- Refactor code for ImagePicker component.

Cleanup AddInventoryDetail component

- the component was difficult to manage, so we create three different files
and properly managed the folder structure for AddInventory for single item.

Tech debt clean for lint issues

Update InventoryList and Table TableComponent

- update inventorylist component to support better layout and styling changes.
We decided to remove redundant code since they are no longer required and just
complicate the system.

- support for bugfix for storage locations not displaying in screen.
The storage locations was not displaying its options and that was because of the redux action
not persisting the storage locations. Since they are tied to the inventories, loading is affected
for both the storage locations and inventories. this makes the inventory page refresh when storage locations
are filled. This is prevelant when the user attempts to Add new inventory.

- backend support for proper asset names.
Updated dataLake to support proper names of the assets, so that they are more in sync of what the application is.

- Tech debt cleanup
Applied tech debt cleanup for a couple of files and kept them into a proper folder structure and removed un-used function
