+++ frontend/src/features/InventoryList/ViewItemDetails/QrCodeGen.jsx
