+++ frontend/src/features/InventoryList/ViewItemDetails/SelectedRowItem.jsx
@@ -1,6 +1,6 @@
 import { Table, TableBody, TableCell, TableRow } from '@mui/material';
-import { BUILD_TABLE_CONSTANTS } from './constants';
-import { EmptyComponent } from '../common/utils';
+import { BUILD_TABLE_CONSTANTS } from '../constants';
+import { EmptyComponent } from '../../common/utils';
 
 const SelectedRowItem = ({ selectedRow, columns }) => {
   if (Object.keys(selectedRow).length === 0) {
